+++ src/dis6/DataReliablePdu.cpp
@@ -1,4 +1,4 @@
-#include <dis6/DataReliablePdu.h>
+#include "dis6/DataReliablePdu.h"
 
 using namespace DIS;
 
@@ -30,9 +30,9 @@ void DataReliablePdu::setRequiredReliabilityService(unsigned char pX) {
   _requiredReliabilityService = pX;
 }
 
-unsigned short DataReliablePdu::getPad1() const { return _pad1; }
+uint16_t DataReliablePdu::getPad1() const { return _pad1; }
 
-void DataReliablePdu::setPad1(unsigned short pX) { _pad1 = pX; }
+void DataReliablePdu::setPad1(uint16_t pX) { _pad1 = pX; }
 
 unsigned char DataReliablePdu::getPad2() const { return _pad2; }
 
@@ -154,12 +154,12 @@ int DataReliablePdu::getMarshalledSize() const {
   marshalSize = marshalSize + 4;  // _numberOfFixedDatumRecords
   marshalSize = marshalSize + 4;  // _numberOfVariableDatumRecords
 
-  for (unsigned long long idx = 0; idx < _fixedDatumRecords.size(); idx++) {
+  for (uint64_t idx = 0; idx < _fixedDatumRecords.size(); idx++) {
     FixedDatum listElement = _fixedDatumRecords[idx];
     marshalSize = marshalSize + listElement.getMarshalledSize();
   }
 
-  for (unsigned long long idx = 0; idx < _variableDatumRecords.size(); idx++) {
+  for (uint64_t idx = 0; idx < _variableDatumRecords.size(); idx++) {
     VariableDatum listElement = _variableDatumRecords[idx];
     marshalSize = marshalSize + listElement.getMarshalledSize();
   }
