+++ src/dis6/EntityStatePdu.cpp
@@ -1,4 +1,4 @@
-#include <dis6/EntityStatePdu.h>
+#include "dis6/EntityStatePdu.h"
 
 using namespace DIS;
 
@@ -228,7 +228,7 @@ int EntityStatePdu::getMarshalledSize() const {
   marshalSize = marshalSize + _marking.getMarshalledSize();  // _marking
   marshalSize = marshalSize + 4;                             // _capabilities
 
-  for (unsigned long long idx = 0; idx < _articulationParameters.size();
+  for (uint64_t idx = 0; idx < _articulationParameters.size();
        idx++) {
     ArticulationParameter listElement = _articulationParameters[idx];
     marshalSize = marshalSize + listElement.getMarshalledSize();
